+++ R/pulseIcon.R
@@ -179,7 +179,7 @@ addPulseMarkers = function(
   }
 
   if (!is.null(clusterOptions))
-    map$dependencies = c(map$dependencies, leaflet::markerClusterDependencies())
+    map$dependencies = c(map$dependencies, leaflet::leafletDependencies$markerCluster())
 
   pts = leaflet::derivePoints(
     data, lng, lat, missing(lng), missing(lat), "addPulseMarkers")