+++ inst/examples/geojson-choropleth.R
@@ -52,5 +52,5 @@ leaf %>%
     color='#ffffff', weight=1, fillOpacity = 0.7,
     highlightOptions =
       highlightOptions(fillOpacity=1, weight=2, opacity=1, color='#000000',
-                        bringToFront=TRUE)
+                        bringToFront=TRUE, sendToBack = TRUE)
  )